Create README.md
ow about Devtoolset container images based on Red Hat Software Collections, that provide a platform for building and running C and C++ applications.And I consider the developers to take charge on Python container images based on Red Hat Software Collections and intended for OpenShift and general usage, that provide a platform for building and running Python applications. Multiply the memory and other system requirements for one server by the number of instances planned for the system.